Move compute util to utils subpackage (former utils submodule is renamed to _coreutils
